+++ cli/src/commands/operation/show.rs
@@ -58,9 +58,9 @@ pub fn cmd_op_show(
     let repo_loader = workspace_command.workspace().repo_loader();
     let settings = workspace_command.settings();
     let op = workspace_command.resolve_single_op(&args.operation)?;
-    let parents: Vec<_> = op.parents().try_collect()?;
-    let parent_op = repo_loader.merge_operations(parents, None)?;
-    let parent_repo = repo_loader.load_at(&parent_op)?;
+    let parent_ops: Vec<_> = op.parents().try_collect()?;
+    let merged_parent_op = repo_loader.merge_operations(parent_ops.clone(), None)?;
+    let parent_repo = repo_loader.load_at(&merged_parent_op)?;
     let repo = repo_loader.load_at(&op)?;
 
     let id_prefix_context = workspace_env.new_id_prefix_context();
@@ -100,6 +100,11 @@ pub fn cmd_op_show(
     let mut formatter = ui.stdout_formatter();
     template.format(&op, formatter.as_mut())?;
 
+    // TODO: Merged repo may have newly rebased commits, which wouldn't exist in
+    // the index. (#4465)
+    if parent_ops.len() > 1 {
+        return Ok(());
+    }
     show_op_diff(
         ui,
         formatter.as_mut(),